MOVEONLY: Move CFeeRate and Amount constants to amount.o
