+++ src/qt/optionsmodel.cpp
@@ -11,6 +11,7 @@
 #include "bitcoinunits.h"
 #include "guiutil.h"
 
+#include "amount.h"
 #include "init.h"
 #include "main.h"
 #include "net.h"
